tests: new detour factory used multiple places
